+++ src/modules/documentExecution/dashboard/widget/interactionsHelpers/IFrameInteractionHelper.ts
@@ -1,36 +1,32 @@
 import { IDashboardDriver, IFrameInteractionSettings, IVariable } from "../../Dashboard"
-import { replaceDriversPlaceholdersByDriverUrlName, replaceVariablesPlaceholdersByVariableName } from "./InteractionLinkHelper"
+import { replaceDriversPlaceholdersByDriverUrlName, replaceFieldPlaceholdersByColumnName, replaceVariablesPlaceholdersByVariableName } from "./InteractionsParserHelper"
 import mainStore from '@/App.store'
 import dashboardStore from '@/modules/documentExecution/dashboard/Dashboard.store'
 import i18n from '@/App.i18n'
-import { columnFieldRegex } from "../../helpers/common/DashboardRegexHelper"
 
 const { t } = i18n.global
 
 
-interface IClickedValue { value: string, type: string }
-
-export const startTableWidgetIFrameInteractions = (clickedValue: IClickedValue, formattedRow: any, iFrameInteractionSettings: IFrameInteractionSettings, dashboardId: string, variables: IVariable[], parentWindow: any) => {
+export const startTableWidgetIFrameInteractions = (formattedRow: any, iFrameInteractionSettings: IFrameInteractionSettings, dashboardId: string, variables: IVariable[], window: any) => {
+    if (!window.parent || window.parent !== window.top) return
     // TODO - FOR TESTING ONLY!
-    parentWindow.addEventListener('message', test)
+    window.parent.addEventListener('message', test)
 
-    if (!parentWindow) return
     const dashStore = dashboardStore()
     const drivers = dashStore.getDashboardDrivers(dashboardId)
     const formattedJSON = getFormattedJSON(iFrameInteractionSettings, variables, drivers, formattedRow, null)
-    sendMessageToParentWindow(parentWindow, formattedJSON)
+    sendMessageToParentWindow(window.parent, formattedJSON)
 }
 
-export const startHTMLAndCustomChartIFrameInteractions = (iframeMessage: string | number, iFrameInteractionSettings: IFrameInteractionSettings, dashboardId: string, variables: IVariable[], parentWindow: any) => {
+export const startHTMLAndCustomChartIFrameInteractions = (iframeMessage: string | number, iFrameInteractionSettings: IFrameInteractionSettings, dashboardId: string, variables: IVariable[], window: any) => {
+    if (!window.parent || window.parent !== window.top) return
     // TODO - FOR TESTING ONLY!
-    parentWindow.addEventListener('message', test)
+    window.parent.addEventListener('message', test)
 
-    if (!parentWindow) return
     const dashStore = dashboardStore()
     const drivers = dashStore.getDashboardDrivers(dashboardId)
     const formattedJSON = getFormattedJSON(iFrameInteractionSettings, variables, drivers, null, iframeMessage)
-    console.log('--------- PARENT WINDOW: ', parentWindow)
-    sendMessageToParentWindow(parentWindow, formattedJSON)
+    sendMessageToParentWindow(window.parent, formattedJSON)
 }
 
 const getFormattedJSON = (iFrameInteractionSettings: IFrameInteractionSettings, variables: IVariable[], drivers: IDashboardDriver[], formattedRow: any, iframeMessage: string | number | null) => {
@@ -62,14 +58,6 @@ const sendMessageToParentWindow = (parentWindow: any, formattedJSON: string) => 
     }, "*")
 }
 
-// TODO - Move to common?
-export const replaceFieldPlaceholdersByColumnName = (originalString: string, formattedRow: any) => {
-    originalString = originalString.replace(columnFieldRegex, (match: string, fieldName: string) => {
-        return formattedRow[fieldName] ? formattedRow[fieldName].value : ''
-    })
-    return originalString
-}
-
 const replacePlaheoldersForHTMLAndCustomChart = (originalString: string, variables: IVariable[], drivers: IDashboardDriver[], iframeMessage: string | number | null) => {
     originalString = replaceVariablesPlaceholdersByVariableName(originalString, variables)
     originalString = replaceDriversPlaceholdersByDriverUrlName(originalString, drivers)
@@ -82,9 +70,7 @@ const replaceFieldPlaceholdersMessagePlaceholder = (originalString, iframeMessag
     return iframeMessage ? originalString.replaceAll(valueRegex, iframeMessage) : ''
 }
 
-
 // TODO - FOR TESTING!
 const test = (event: any) => {
-    // console.log('----- TEST: ', event)
-    //console.log('----- TEST MESSAGE: ', event.data)
+    console.log('----- TEST MESSAGE: ', event.data)
 }
